Refactor form on login page to its own component
